+++ services/number_setting.go
@@ -8,8 +8,8 @@ import (
 
 type INumberSettingService interface {
 	Create(numberSetting models.NumberSetting, language string) (models.NumberSetting, error)
-	Find(per, page uint, filter map[string]interface{}, selectKeys []string, order []string) (numberSettings []models.ResultNumberSetting,
-		total uint, err error)
+	Find(per, page int, filter map[string]interface{}, selectKeys []string, order []string) (numberSettings []models.ResultNumberSetting,
+		total int64, err error)
 	Delete(id uint) error
 }
 
@@ -21,8 +21,8 @@ func (n NumberSettingService) Delete(id uint) error {
 	return n.repo.Delete(id)
 }
 
-func (n NumberSettingService) Find(per, page uint, filter map[string]interface{}, selectKeys []string, order []string) (numberSettings []models.ResultNumberSetting,
-	total uint, err error) {
+func (n NumberSettingService) Find(per, page int, filter map[string]interface{}, selectKeys []string, order []string) (numberSettings []models.ResultNumberSetting,
+	total int64, err error) {
 	return n.repo.Find(per, page, filter, selectKeys, order, true)
 }
